harden kernels + codec robustness (code-review follow-up)
  - ops: int64 指针偏移加固 5 个 kernel（对齐 flash_attn，防 M·V/M·N ≥ 2^31 溢出）
  - fused_rms_norm: 统一默认 eps 1e-10→1e-6（与 RMSNorm 一致）
  - verify_lossless: 截断 MDLC 容器抛 ValueError（修 /api/decode 500→400）
  - igpt: docstring 注明 head 前无 final RMSNorm 为刻意偏差
  - imagenet64 config: 修正 warmup_epochs=1 误导注释

+++ src/mdlic/ops/fused_add_rms_norm.py
@@ -53,7 +53,7 @@ def _fused_add_rms_norm_fwd_kernel(
     每个 program 处理一行。
     y_i = residual_i + w_i * sublayer_i * rrms
     """
-    row = tl.program_id(0)
+    row = tl.program_id(0).to(tl.int64)   # int64：避免 M·N ≥ 2^31 时 row*stride 指针偏移溢出（对齐 flash_attn.py）
     cols = tl.arange(0, BLOCK_N)
     mask = cols < N
 
@@ -100,7 +100,7 @@ def _fused_add_rms_norm_bwd_kernel(
       其中 sublayer_hat = sublayer * rrms
     dw_row = dy * sublayer_hat
     """
-    row = tl.program_id(0)
+    row = tl.program_id(0).to(tl.int64)   # int64：避免 M·N ≥ 2^31 时 row*stride 指针偏移溢出（对齐 flash_attn.py）
     cols = tl.arange(0, BLOCK_N)
     mask = cols < N
